requirements.txt
streamlit==1.33.0
python-dotenv==1.0.1

# LangChain
langchain==0.2.14
langchain-core==0.2.36
langchain-community==0.2.12
langchain-openai==0.1.22
langchain-text-splitters==0.2.2

# Vector DB
faiss-cpu==1.8.0

# Embeddings
sentence-transformers==2.2.2

# Reranking
rank-bm25==0.2.2

# PDF processing
unstructured==0.14.10
pdfplumber==0.11.4
pypdf==4.2.0

# Utilities
numpy
scikit-learn
tqdm

# OpenAI compatible
openai==1.40.6